updated repeat command timers to use threads, the threads finish immediately when user repeat timers are set to 0
