+++ ChatBotNet5/Models/TimerCommand.cs
@@ -7,9 +7,24 @@ namespace ChatBot_Net5.Models
     [DebuggerDisplay("Command={Command}, RepeatTime={RepeatTime}, NextRun={NextRun}")]
     internal class TimerCommand : IComparable<TimerCommand>, IEquatable<TimerCommand>
     {
+        /// <summary>
+        /// The Command represented in this command repeater
+        /// </summary>
         internal string Command { get; set; }
+
+        /// <summary>
+        /// The amount of seconds between each run.
+        /// </summary>
         internal int RepeatTime { get; set; }
+
+        /// <summary>
+        /// The DateTime of the next repeat timer run for this event.
+        /// </summary>
         internal DateTime NextRun { get; set; }
+
+        /// <summary>
+        /// The list of categories applying to this command, run command if category matches current stream or "All" category designation.
+        /// </summary>
         internal List<string> CategoryList { get; } = new();
 
         internal TimerCommand(Tuple<string, int, string[]> ComRepeat, double TimeDilute)
@@ -24,7 +39,7 @@ namespace ChatBot_Net5.Models
 
         internal bool CheckFireTime() => DateTime.Now > NextRun;
 
-        public int CompareTo(TimerCommand obj) => RepeatTime.CompareTo(obj.RepeatTime);
+        public int CompareTo(TimerCommand obj) => Command.CompareTo(obj.Command);
 
         public bool Equals(TimerCommand other) => Command == other.Command;
