+++ ChatBotNet5/Systems/CommandSystem.cs
@@ -11,6 +11,7 @@ using System.ComponentModel;
 using System.Globalization;
 using System.Linq;
 using System.Threading;
+using System.Threading.Tasks;
 
 using TwitchLib.Client.Models;
 
@@ -73,6 +74,7 @@ namespace ChatBot_Net5.Systems
 
             const int ChatCount = 20;
             const int ViewerCount = 10;
+            const int ThreadSleep = 5000;
 
             double CheckDilute()
             {
@@ -110,62 +112,75 @@ namespace ChatBot_Net5.Systems
                 return 1.0;
             }
 
-            double DiluteTime = CheckDilute();
-
-
-            foreach (Tuple<string, int, string[]> Timers in datamanager.GetTimerCommands())
+            void RepeatCmd(ref TimerCommand cmd)
             {
-                if (Timers.Item3.Contains(StatData.Category) || Timers.Item3.Contains(LocalizedMsgSystem.GetVar(Msg.MsgAllCateogry)))
+                int repeat = 0;
+                bool InCategory = false;
+
+                lock (cmd) // lock the cmd because it's referenced in other threads
                 {
-                    RepeatList.Add(new(Timers, DiluteTime));
+                    repeat = cmd.RepeatTime;
+                    InCategory = (cmd.CategoryList.Contains(StatData.Category) || cmd.CategoryList.Contains(LocalizedMsgSystem.GetVar(Msg.MsgAllCateogry)));
+                }
+
+                while (
+                    OptionFlags.ProcessOps
+                    && repeat != 0
+                    && InCategory)
+                {
+                    if ((cmd.NextRun - DateTime.Now).Seconds <= 0)
+                    {
+                        OnRepeatEventOccured?.Invoke(this, new TimerCommandsEventArgs() { Message = PerformCommand(cmd.Command, BotUserName, null, true) });
+                        lock (cmd)
+                        {
+                            cmd.UpdateTime(CheckDilute());
+                        }
+                    }
+                    Thread.Sleep(ThreadSleep);
+
+// TODO: check the repeat time is being updated so command can be stopped from repeating
+                    lock (cmd) // lock the cmd because it's referenced in other threads
+                    {
+                        repeat = cmd.RepeatTime;
+                        InCategory = (cmd.CategoryList.Contains(StatData.Category) || cmd.CategoryList.Contains(LocalizedMsgSystem.GetVar(Msg.MsgAllCateogry)));
+                    }
                 }
             }
 
+            double DiluteTime = CheckDilute();
+
             while (OptionFlags.ProcessOps)
             {
-                DiluteTime = CheckDilute();
-
-                foreach (TimerCommand timer in RepeatList)
-                {
-                    if (timer.CheckFireTime() && (timer.CategoryList.Contains(StatData.Category) || timer.CategoryList.Contains(LocalizedMsgSystem.GetVar(Msg.MsgAllCateogry))))
-                    {
-                        timer.UpdateTime(DiluteTime);
-                        string output = PerformCommand(timer.Command, BotUserName, null, true);
-
-                        OnRepeatEventOccured?.Invoke(this, new TimerCommandsEventArgs() { Message = output });
-                    }
-                }
-
-                // check if any commands are added to the repeat timers
                 foreach (Tuple<string, int, string[]> Timers in datamanager.GetTimerCommands())
                 {
-                    TimerCommand command = new(Timers, DiluteTime);
-                    if (command.CategoryList.Contains(StatData.Category) || command.CategoryList.Contains(LocalizedMsgSystem.GetVar(Msg.MsgAllCateogry)))
+                    if (Timers.Item3.Contains(StatData.Category) || Timers.Item3.Contains(LocalizedMsgSystem.GetVar(Msg.MsgAllCateogry)))
                     {
-                        if (!RepeatList.Contains(command))
+                        TimerCommand item = new(Timers, DiluteTime);
+                        if (!RepeatList.Contains(item))
                         {
-                            RepeatList.Add(command);
-                            string output = PerformCommand(command.Command, BotUserName, null, true);
-
-                            OnRepeatEventOccured?.Invoke(this, new TimerCommandsEventArgs() { Message = output });
+                            RepeatList.Add(item);
+                            new Thread(new ThreadStart(()=>RepeatCmd(ref item))).Start();
                         }
                         else
                         {
-                            TimerCommand update = RepeatList.Find((a) => a.Command == command.Command);
-                            update.RepeatTime = command.RepeatTime;
+                            TimerCommand Listcmd = RepeatList.Find((f) => f.Command == item.Command);
+
+                            if (item.RepeatTime == 0)
+                            {
+                                RepeatList.Remove(Listcmd);
+                            }
+                            else
+                            {
+                                lock (Listcmd)
+                                {
+                                    Listcmd.RepeatTime = item.RepeatTime; // update repeat time, the task will update time after it runs one loop of repeat
+                                }
+                            }
                         }
                     }
                 }
 
-                for (int x = RepeatList.Count - 1; x >= 0 && RepeatList.Count > 0; x--)
-                {
-                    if (RepeatList[x].RepeatTime == 0 || !(RepeatList[x].CategoryList.Contains(LocalizedMsgSystem.GetVar(Msg.MsgAllCateogry)) || RepeatList[x].CategoryList.Contains(StatData.Category)))
-                    {
-                        RepeatList.RemoveAt(x);
-                    } 
-                }
-
-                Thread.Sleep(5000); // wait for awhile before checking commands again
+                Thread.Sleep((int)(ThreadSleep*1.5)); // wait for awhile before checking commands again
             }
         }
 
@@ -300,9 +315,6 @@ namespace ChatBot_Net5.Systems
                 return (OptionFlags.MsgPerComMe && CommData.AddMe ? "/me " : "") + VariableParser.ParseReplace(CommData?.Message ?? LocalizedMsgSystem.GetVar(ChatBotExceptions.ExceptionKeyNotFound), datavalues);
             }
         }
-    
-     
-
 
         /*
          if (command == "addcommand")
